requirements.txt
cachelib==0.10.2
Flask==2.3.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.5
Flask-Session==0.5.0 
Jinja2==3.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python_dateutil==2.8.2
SQLAlchemy==2.0.15
Werkzeug==2.3.4

